update user invoice
